% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleListFunction.R
\name{createSampleList}
\alias{createSampleList}
\title{Creating sample list to run mcapply parallel}
\usage{
createSampleList(samples, reads_folder, column, program, libraryType, fileType, fromSTAR = FALSE)
}
\arguments{
\item{samples}{\code{Character.} The filename of the sample file. Default samples.txt.}

\item{reads_folder}{\code{Character.} Directory where the raw sequence data is stored. Default 00-Fastq.}

\item{column}{\code{Character.} Column name from the sample sheet to use as read folder names. Default SAMPLE_ID}

\item{program}{\code{Character.} The program to be used. Available 'trimmomatic', 'star', 'Hisat2', 'htseq', 'featurecount', 'bowtie', 'samtools' or 'bwa'}

\item{libraryType}{\code{Character.} The library type to use. Available: 'pairEnd' or 'singleEnd'.}

\item{fileType}{\code{Character.} The file type to use. Available: 'bam' or 'sam'.}

\item{fromSTAR}{\code{Logical.} Wheter to use samples from STAR in FeatureCouts}
}
\description{
Function to create sample list to run mcapply parallel
}
\details{
December 2021
}
\author{
Haniel Cedraz
}
