% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSamplesFileFunction.R
\name{loadSamplesFile}
\alias{loadSamplesFile}
\title{Loading sample file}
\usage{
loadSamplesFile(file, reads_folder, column = "SAMPLE_ID", libraryType = "pairEnd")
}
\arguments{
\item{file}{\code{Character.} The filename of the sample file. Default samples.txt.}

\item{reads_folder}{\code{Character.} Directory where the raw sequence data is stored. Default 00-Fastq.}

\item{column}{\code{Character.} Column name from the sample sheet to use as read folder names. Default SAMPLE_ID}

\item{libraryType}{\code{Character.} The library type to use. Available: 'pairEnd' or 'singleEnd'.}
}
\description{
Function to load the sample file
}
\details{
December 2021
}
\author{
Haniel Cedraz
}
