% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCoreFunction.R
\name{prepareCore}
\alias{prepareCore}
\title{prepare number of cores to use}
\usage{
prepareCore(nThreads = 8)
}
\arguments{
\item{nThreads}{\code{integer.} Number of processors to use. Default 8.}
}
\description{
prepare number of cores to use
}
\details{
December 2021
}
\author{
Haniel Cedraz
}
